requirements.txt
pypdf>=4.2.0
pdfplumber>=0.11.0
python-docx>=1.1.0
openpyxl>=3.1.0

